feat(test): add test harness for integration tests
- Create test/harness module with MockAtom trait and repo initialization utilities
- Add mock_with_deps method for creating atoms with dependencies in tests
- Refactor publish/git tests to use shared harness infrastructure
- Make AtomWriter and WriteDeps pub(crate) for test access
- Move init_tracing and init_repo_and_remote to harness for reuse
