+++ crates/atom/src/test/mod.rs
@@ -1,0 +1,5 @@
+//! # Test Utilities
+//!
+//! Shared test infrastructure for the atom crate.
+
+pub mod harness;
